データのばらつきを抑制するスケール変換器の追加，ソースの整形

+++ java/src/net/coolblossom/lycee/machinelearning/classification/OnlineLearning.java
@@ -2,6 +2,7 @@ package net.coolblossom.lycee.machinelearning.classification;
 
 /**
  * オンライン学習
+ * 
  * @author ryouka0122@github
  *
  */
@@ -9,7 +10,9 @@ public abstract class OnlineLearning extends AbstractClassifier {
 
 	/**
 	 * コンストラクタ
-	 * @param dataDimension データ次数（バイアス項含む）
+	 * 
+	 * @param dataDimension
+	 *            データ次数（バイアス項含む）
 	 */
 	public OnlineLearning(int dataDimension) {
 		super(dataDimension);
@@ -17,9 +20,12 @@ public abstract class OnlineLearning extends AbstractClassifier {
 
 	/**
 	 * 更新
-	 * @param correctLabel 正しいラベル
-	 * @param data 観測データ
+	 * 
+	 * @param correctLabel
+	 *            正しいラベル
+	 * @param data
+	 *            観測データ
 	 */
-	abstract public void refine(int correctLabel, double[] data);
+	abstract public void refine(double correctLabel, double[] data);
 
 }
