+++ java/src/net/coolblossom/lycee/machinelearning/classification/online/PossiveAggressiveAlgorithm.java
@@ -5,6 +5,7 @@ import net.coolblossom.lycee.machinelearning.classification.loss.LossLogic;
 
 /**
  * Possive-Aggressive Algorithm
+ * 
  * @author ryouka0122@github
  *
  */
@@ -21,9 +22,13 @@ public class PossiveAggressiveAlgorithm extends OnlineLearning {
 
 	/**
 	 * コンストラクタ
-	 * @param dimension 観測データ項数（定数項含む）
-	 * @param type PAアルゴリズムの種類
-	 * @param C 正則化項の重みパラメータ
+	 * 
+	 * @param dimension
+	 *            観測データ項数（定数項含む）
+	 * @param type
+	 *            PAアルゴリズムの種類
+	 * @param C
+	 *            正則化項の重みパラメータ
 	 */
 	public PossiveAggressiveAlgorithm(int dimension, PAType type, double C) {
 		super(dimension);
@@ -33,16 +38,17 @@ public class PossiveAggressiveAlgorithm extends OnlineLearning {
 	}
 
 	@Override
-	public void refine(int correctLabel, double[] data) {
+	public void refine(double correctLabel, double[] data) {
 		// 損失の計算
 		double lossValue = loss.calc(parameters, correctLabel, data);
-		if(lossValue==0.0) return;
+		if (lossValue == 0.0)
+			return;
 
 		// 更新率の計算
 		double K = type.calc(C, lossValue, data);
 
 		// 更新
-		for(int i=0 ; i<parameters.length ; i++) {
+		for (int i = 0; i < parameters.length; i++) {
 			parameters[i] += K * correctLabel * data[i];
 		}
 	}
